fixed branch recycler

+++ app/src/main/java/com/tubespbdandroid/majika/data/Branch.kt
@@ -1,10 +1,12 @@
 package com.tubespbdandroid.majika.data
 
-data class Branch(val name:String,
-                  val popularFood:String,
-                  val contactPerson:String,
-                  val address:String,
-                  val phone:String,
+data class Branch(
+    val name:String,
+    val popularFood:String,
+    val contactPerson:String,
+    val address:String,
+    val phone:String,
 
-                  val longitude:Float,
-                  val latitude:Float)
+    val longitude: Double,
+    val latitude: Double
+)
